+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Changed
 - Denote when in “read only” mode. ([#22](https://github.com/doublesecretagency/craft-inventory/issues/22))
 
+### Fixed
+- Improved handling of field deletion.
+
 ## 3.1.0 - 2023-06-30
 
 ### Fixed